+++ airflow/ui/openapi-gen/requests/types.gen.ts
@@ -628,6 +628,8 @@ export type DAGRunsBatchBody = {
   page_limit?: number;
   dag_ids?: Array<string> | null;
   states?: Array<DagRunState | null> | null;
+  run_after_gte?: string | null;
+  run_after_lte?: string | null;
   logical_date_gte?: string | null;
   logical_date_lte?: string | null;
   start_date_gte?: string | null;
@@ -820,6 +822,7 @@ export type DagVersionResponse = {
   bundle_name: string;
   bundle_version: string | null;
   created_at: string;
+  readonly bundle_url: string | null;
 };
 
 /**
@@ -906,8 +909,10 @@ export type GridDAGRunwithTIs = {
   queued_at: string | null;
   start_date: string | null;
   end_date: string | null;
+  run_after: string;
   state: DagRunState;
   run_type: DagRunType;
+  logical_date: string | null;
   data_interval_start: string | null;
   data_interval_end: string | null;
   version_number: number | null;
@@ -1783,6 +1788,8 @@ export type GridDataData = {
   offset?: number;
   orderBy?: string;
   root?: string | null;
+  runAfterGte?: string | null;
+  runAfterLte?: string | null;
   runType?: Array<string>;
   state?: Array<string>;
 };
@@ -1884,6 +1891,8 @@ export type GetDagRunsData = {
   logicalDateLte?: string | null;
   offset?: number;
   orderBy?: string;
+  runAfterGte?: string | null;
+  runAfterLte?: string | null;
   startDateGte?: string | null;
   startDateLte?: string | null;
   state?: Array<string>;