+++ providers/google/tests/system/google/cloud/cloud_memorystore/example_cloud_memorystore_memcached.py
@@ -29,6 +29,8 @@ from __future__ import annotations
 import os
 from datetime import datetime
 
+from google.protobuf.field_mask_pb2 import FieldMask
+
 from airflow.models.dag import DAG
 from airflow.providers.google.cloud.operators.cloud_memorystore import (
     CloudMemorystoreMemcachedApplyParametersOperator,
@@ -41,7 +43,6 @@ from airflow.providers.google.cloud.operators.cloud_memorystore import (
 )
 from airflow.providers.standard.operators.bash import BashOperator
 from airflow.utils.trigger_rule import TriggerRule
-from google.protobuf.field_mask_pb2 import FieldMask
 
 ENV_ID = os.environ.get("SYSTEM_TESTS_ENV_ID", "default")
 PROJECT_ID = os.environ.get("SYSTEM_TESTS_GCP_PROJECT", "google_project_id")